+++ src/modules/populates/reports.js
@@ -1,0 +1,10 @@
+const reportPopulateOption = [
+  {
+    path: "reportedId",
+    select: "_id id name nickname profileImageUrl",
+  },
+];
+
+module.exports = {
+  reportPopulateOption,
+};
